+++ packages/experiments/src/components/CollapsibleSection/CollapsibleSectionTitle.view.tsx
@@ -1,6 +1,7 @@
 /** @jsx withSlots */
 import { Text } from '../../Text';
 import { Icon } from 'office-ui-fabric-react/lib/Icon';
+import { getNativeProps, buttonProperties } from 'office-ui-fabric-react/lib/Utilities';
 import { withSlots, getSlots } from '../../Foundation';
 import {
   ICollapsibleSectionTitleComponent,
@@ -15,8 +16,10 @@ export const CollapsibleSectionTitleView: ICollapsibleSectionTitleComponent['vie
     text: Text
   });
 
+  const buttonProps = getNativeProps(props, buttonProperties);
+
   return (
-    <Slots.root ref={props.focusElementRef} onClick={props.onClick} onKeyDown={props.onKeyDown}>
+    <Slots.root {...buttonProps} ref={props.focusElementRef}>
       {!props.chevronDisabled && <Slots.chevron iconName="ChevronDown" />}
       <Slots.text />
     </Slots.root>